+++ downloader/Downloader.py
@@ -1,12 +1,10 @@
 
 class Downloader:
-    def __init__(self, collection, logger, temp_dir, cache_dir, user, password) -> None:
-        self.collection = collection
+    def __init__(self, dataset, logger, temp_dir, cache_dir) -> None:
+        self.dataset = dataset
         self.logger = logger
         self.temp_dir = temp_dir
         self.cache_dir = cache_dir
-        self.user = user
-        self.password = password
 
 def run(self):
     raise Exception(f"Run method not implemented")