+++ downloader/Manual.py
@@ -6,8 +6,8 @@ from utils.utils import run_subprocess
 
 
 class Manual(Downloader):
-    def __init__(self, collection, logger, temp_dir, cache_dir, user=None, password=None):
-        super(Manual, self).__init__(collection=collection, logger=logger, temp_dir=temp_dir, cache_dir=cache_dir, user=user, password=password)
+    def __init__(self, collection, logger, temp_dir, cache_dir, credentials=None):
+        super(Manual, self).__init__(dataset=collection, logger=logger, temp_dir=temp_dir, cache_dir=cache_dir)
 
     
     def run(self):